pyproject.toml
[tool.poetry]
name = "weblate-batch-scripts"
version = "0.1.0"
description = ""
authors = ["samzong.lu <samzong.lu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.78.0"
uvicorn = "0.18.2"
requests = "2.28.1"
anyio = "3.6.1"
certifi = "2022.6.15"
charset-normalizer = "2.1.0"
click = "8.1.3"
h11 = "0.13.0"
idna = "3.3"
pip = "22.2"
pydantic = "1.9.1"
setuptools = "62.6.0"
sniffio = "1.2.0"
starlette = "0.19.1"
typing-extensions = "4.3.0"
urllib3 = "1.26.10"
wheel = "0.37.1"
wlc = "^1.13"
python-gitlab = "^3.12.0"
gitpython = "^3.1.29"
jinja2 = "^3.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
